requirements.txt
coverage>=6.0.1
gensim==3.8.3
numpy>=1.19.5
onnx>=1.10.1
onnxruntime>=1.9.0
pre-commit>=2.15.0
PyYAML>=5.4.1
scikit-learn>=0.24.2
torch>=1.9.1
tqdm>=4.62.3
